+++ CHANGELOG.md
@@ -27,6 +27,7 @@
  * configure `fastq_sceen` in the uap config (#160)
  * `count_rRNA` doese not fail silently for uncomressed input (#156)
  * detect uncommitded changes even if they are staged
+ * no default seed for adapterRemoval (#149)
 
 **Features**
  * use relative paths; uap output can now be moved (#115)